Move disaggregation files to internal
