+++ tt_metal/impl/internal/disaggregation/umd_dram_reader.cpp
@@ -12,7 +12,7 @@
 // sidecar), so no MetalContext / ControlPlane is touched (either would start_device the chips).
 // ---------------------------------------------------------------------------
 
-#include "tt_metal/impl/experimental/disaggregation/noc_addr.hpp"
+#include "tt_metal/impl/internal/disaggregation/noc_addr.hpp"
 
 #include <tt_stl/assert.hpp>
 
@@ -37,7 +37,7 @@
 #include <umd/device/types/core_coordinates.hpp>
 #pragma GCC diagnostic pop
 
-namespace tt::tt_metal::experimental::disaggregation {
+namespace tt::tt_metal::internal::disaggregation {
 
 namespace {
 
@@ -136,4 +136,4 @@ std::vector<uint8_t> read_dram_umd(uint64_t unique_id, uint64_t noc_addr, uint32
     return buf;
 }
 
-}  // namespace tt::tt_metal::experimental::disaggregation
+}  // namespace tt::tt_metal::internal::disaggregation
